requirements.txt
spotipy
google-cloud-bigquery
google-cloud-secret-manager
python-dotenv
requests